internal: move diagnostics to a new crate
