+++ crates/ide_diagnostics/Cargo.toml
@@ -10,3 +10,21 @@ edition = "2018"
 doctest = false
 
 [dependencies]
+cov-mark = "2.0.0-pre.1"
+itertools = "0.10.0"
+rustc-hash = "1.1.0"
+either = "1.5.3"
+
+profile = { path = "../profile", version = "0.0.0" }
+stdx = { path = "../stdx", version = "0.0.0" }
+syntax = { path = "../syntax", version = "0.0.0" }
+text_edit = { path = "../text_edit", version = "0.0.0" }
+cfg = { path = "../cfg", version = "0.0.0" }
+hir = { path = "../hir", version = "0.0.0" }
+ide_db = { path = "../ide_db", version = "0.0.0" }
+ide_assists = { path = "../ide_assists", version = "0.0.0" }
+
+[dev-dependencies]
+expect-test = "1.1"
+
+test_utils = { path = "../test_utils" }
