+++ crates/ide_diagnostics/src/missing_match_arms.rs
@@ -1,6 +1,6 @@
 use hir::InFile;
 
-use crate::diagnostics::{Diagnostic, DiagnosticsContext};
+use crate::{Diagnostic, DiagnosticsContext};
 
 // Diagnostic: missing-match-arm
 //
@@ -18,11 +18,11 @@ pub(super) fn missing_match_arms(
 
 #[cfg(test)]
 pub(super) mod tests {
-    use crate::diagnostics::tests::check_diagnostics;
+    use crate::tests::check_diagnostics;
 
     fn check_diagnostics_no_bails(ra_fixture: &str) {
         cov_mark::check_count!(validate_match_bailed_out, 0);
-        crate::diagnostics::tests::check_diagnostics(ra_fixture)
+        crate::tests::check_diagnostics(ra_fixture)
     }
 
     #[test]
