+++ crates/ide_diagnostics/src/unimplemented_builtin_macro.rs
@@ -1,7 +1,4 @@
-use crate::{
-    diagnostics::{Diagnostic, DiagnosticsContext},
-    Severity,
-};
+use crate::{Diagnostic, DiagnosticsContext, Severity};
 
 // Diagnostic: unimplemented-builtin-macro
 //
