+++ crates/ide_diagnostics/src/unresolved_macro_call.rs
@@ -1,7 +1,7 @@
 use hir::{db::AstDatabase, InFile};
 use syntax::{AstNode, SyntaxNodePtr};
 
-use crate::diagnostics::{Diagnostic, DiagnosticsContext};
+use crate::{Diagnostic, DiagnosticsContext};
 
 // Diagnostic: unresolved-macro-call
 //
@@ -32,7 +32,7 @@ pub(super) fn unresolved_macro_call(
 
 #[cfg(test)]
 mod tests {
-    use crate::diagnostics::tests::check_diagnostics;
+    use crate::tests::check_diagnostics;
 
     #[test]
     fn unresolved_macro_diag() {
